+++ JPetLargeBarrelExtensions/TimeSyncDeltas.cpp
@@ -24,8 +24,8 @@ Synchronization::Synchronization(const shared_ptr < AbstractBarrelMapping > map,
 Synchronization::~Synchronization(){}
 const SynchroStrip Synchronization::get_times(const JPetHit & hit) const{
     const auto& offsets=f_offsets->operator[](f_mapping->getStripPos(hit.getBarrelSlot()));
-    map<int,double> lead_times_A = hit.getSignalA().getRecoSignal().getRawSignal().getTimesVsThresholdNumber(JPetSigCh::Leading);
-    map<int,double> lead_times_B = hit.getSignalB().getRecoSignal().getRawSignal().getTimesVsThresholdNumber(JPetSigCh::Leading);
+    map<unsigned int,float> lead_times_A = hit.getSignalA().getRecoSignal().getRawSignal().getTimesVsThresholdNumber(JPetSigCh::Leading);
+    map<unsigned int,float> lead_times_B = hit.getSignalB().getRecoSignal().getRawSignal().getTimesVsThresholdNumber(JPetSigCh::Leading);
     vector<double> A,B;
     for(size_t thr=1;thr<=4;thr++){
 	A.push_back(lead_times_A[thr]);
